refactor blog portable text
