+++ apps/shop-bcd/src/app/[lang]/blog/[slug]/page.tsx
@@ -1,5 +1,6 @@
 import { notFound } from "next/navigation";
 import { fetchPostBySlug } from "@acme/sanity";
+import { BlogPortableText } from "@/components/blog/BlogPortableText";
 import shop from "../../../../../shop.json";
 
 export default async function BlogPostPage({
@@ -14,10 +15,8 @@ export default async function BlogPostPage({
       <h1 className="text-2xl font-bold">{post.title}</h1>
       {post.excerpt && <p className="text-muted">{post.excerpt}</p>}
       {Array.isArray(post.body) ? (
-        <div className="space-y-2">
-          {(post.body as any[]).map((b, i) => (
-            <p key={i}>{b.children?.map((c: any) => c.text).join("")}</p>
-          ))}
+        <div className="space-y-4">
+          <BlogPortableText value={post.body} />
         </div>
       ) : null}
     </article>
